requirements.txt
aniso8601==3.0.2
certifi==2018.8.24
dj-database-url==0.5.0
Django==2.1.1
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.8.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
promise==2.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-decouple==3.1
pytz==2018.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.6
whitenoise==4.1
